[#126] Change FilePond for Uppy as we need S3 policy to upload
Uppy has support for uploading to AWS S3. In the previous code the
image was uploaded directly after selection. We can't do that anymore.
We need to request a policy to upload to S3, so we wait until the
_Submit_ process to upload all files.

There is a single uppy instance that needs to be passed around, from
Home to child components.