[#134] Fix table definition using md5() for unique value
We want to make sure we don't store the same `state` value more than
once.  Usually this is done by marking a column as `UNIQUE` which
creates a btree index. This indexes have a limitation on the size
leading to an error:

    index row requires <large> bytes, maximum size is 8191

To solve this you usually index using a hash function on the content.
We have done the same in other places, e.g.

https://github.com/akvo/akvo-unified-log/blob/870d27531f24549ebdb6f049bec34642f081248f/resources/migrations/001-jsonb-unique-constraint.up.sql#L10