Source not needed inside the container
The containers are not deployed at all in production, so there is no need to copy the sources inside the container, as Docker-Compose mounts the "live" directory of the host in the same location.

Doing a CP can take several minutes if you are restarting the containers, and docker layer is not going to help if you have touched any of the files in the project.